Clean up borrowing of guest VMSA structures
This is safe because the guest VMSA reference is not held in the PerCpu
object, but in a PerCpuShared object which is never mutably borrowed in
any case.  To make VMSA references more robust, VMSA contents are always
referenced while the lock guard on the guest VMSA still exists, ensuring
that the VMSA cannot be remapped while any borrow on its contents is
still alive.

Signed-off-by: Jon Lange <jlange@microsoft.com>